LayoutTests/fast/dom/DeviceOrientation/no-synchronous-events-expected.txt

Tests that events are never fired sycnhronously from a call to window.addEventListener().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS hasAddEventListenerReturned is true
PASS successfullyParsed is true

TEST COMPLETE
